fresh connect on every api call
